+++ src/types/IFunction.ts
@@ -2,5 +2,6 @@ import { PartialAlert } from './config/ILibratoAlert';
 
 export interface IFunction {
   name: string;
+  environment?: { [name: string]: string };
   libratoAlerts?: (string | PartialAlert)[];
 }